move FacetValue into a dedicated module
https://coveord.atlassian.net/browse/JSUI-2944
