+++ src/ui/Facet/Facet.ts
@@ -57,7 +57,8 @@ import { FacetSort } from './FacetSort';
 import { FacetSortCriterion } from './FacetSortCriterion';
 import { FacetUtils } from './FacetUtils';
 import { FacetValueElement } from './FacetValueElement';
-import { FacetValue, FacetValues } from './FacetValues';
+import { FacetValue } from './FacetValue';
+import { FacetValues } from './FacetValues';
 import { FacetValuesList } from './FacetValuesList';
 import { FacetValuesOrder } from './FacetValuesOrder';
 import { OmniboxValueElement } from './OmniboxValueElement';
