+++ src/ui/Facet/FacetValueElement.ts
@@ -1,7 +1,7 @@
 /// <reference path="Facet.ts" />
 
 import { Facet } from './Facet';
-import { FacetValue } from './FacetValues';
+import { FacetValue } from './FacetValue';
 import { ValueElement } from './ValueElement';
 
 export interface IFacetValueElementKlass {
