+++ src/ui/Facet/OmniboxValueElement.ts
@@ -1,7 +1,7 @@
 /// <reference path="Facet.ts" />
 
 import { Facet } from './Facet';
-import { FacetValue } from './FacetValues';
+import { FacetValue } from './FacetValue';
 import { IPopulateOmniboxObject } from '../Omnibox/OmniboxInterface';
 import { ValueElement } from './ValueElement';
 import { IAnalyticsActionCause } from '../Analytics/AnalyticsActionListMeta';
