+++ src/ui/HierarchicalFacet/HierarchicalFacetValueElement.ts
@@ -2,7 +2,7 @@
 
 import { FacetValueElement } from '../Facet/FacetValueElement';
 import { HierarchicalFacet } from './HierarchicalFacet';
-import { FacetValue } from '../Facet/FacetValues';
+import { FacetValue } from '../Facet/FacetValue';
 
 export class HierarchicalFacetValueElement extends FacetValueElement {
   constructor(public facet: HierarchicalFacet, public facetValue: FacetValue, public keepDisplayedValueNextTime: boolean) {
